python/pyproject.toml
[project]
name = "python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
    "pdfplumber>=0.11.7",
    "ruff>=0.12.2",
    "pytesseract>=0.3.10",
    "pdf2image>=1.17.0",
    "Pillow>=10.2.0",
    "opencv-python>=4.9.0",
    "numpy>=1.26.0",
    "faster-whisper>=1.0.1",
    "pydub>=0.25.1",
    "tqdm>=4.66.0",
    "noisereduce>=3.0.3",
    "httpx>=0.27.0",
    "playwright>=1.47.0",
    "openai>=1.51.0",
    "python-dotenv>=1.0.1",
    "google-cloud-speech>=2.27.0",
    "google-cloud-storage>=2.18.0",
    "typer>=0.12.3",
    "pydantic>=2.8.2",
    "pydantic-settings>=2.4.0",
    "beautifulsoup4>=4.12.2",
    "lxml>=5.2.2",
    "rich>=13.7.1",
    "pyyaml>=6.0.1",
]

[project.scripts]
site-inspector = "site_inspector.cli:app"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
